requirements.txt
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.4
django-hstore==1.4.2
django-storages==1.6.5
djangorestframework==3.6.4
djangorestframework-hstore==1.3
django-grappelli==2.13.1
django-filter==2.1.0
requests==2.22.0
celery==4.3.0